add prefs subcommand to organize
This allows the PrefsCommands suite of commands to be used directly on a
freshly-instantiated Organize object without going through dbus, as was already
possible with the PeerCommands.

This is mostly useful for debugging and development, but it also provides a way
to edit the prefs file without dbus or an organize daemon running. This should
not actually be used to set preferences while an organize daemon is running,
however, because the daemon will not re-read its state file and it will clobber
any changes made externally.

Unlike the "peer" command on organize, the name "prefs" is masking the
attribute-reading command which was previously automatically present on the
organize object. (The output of that command had the same output as the prefs
command without arguments.)