Adds golang implementation of vula
This adds a highly experimental second implementation of vula in go. It is not yet reviewed for correctness of the vula protocol or security considerations beyond basic interoperability.

It was done by @domini and @phipus as part of the Application Security class at Bern University of Applied Sciences in the spring semester of 2025.

The goal of this implementation is to create a portable version of vula that can be used as a drop-in replacement of the python version. Currently, the main logic of the commands `discover`, `publish` and `organize` is implemented. One can replace the vula executable in the systemd unit files with this vula go binary. To navigate the codebase, start at `internal/cmd/`. For example, open `internal/cmd/organize.go` and inspect the `Run` function, which is the entry point of the command.

The commands `discover`, `publish` and `organize` are commands that start services which export dbus interfaces. When they are started, one can use the python client to interact with them. For an example of a regular command implementation, see `internal/cmd/status.go`.

To build the binary, run `go run ./cmd/vula`. To build in a container with podman, run `./helper build_simple`. To run the tests, run `make gotest` in the `podman` directory or `./helper test`. Then, the binary can be used from the systemd unit files or directly (e.g. `./vula status`).